small refactoring - methods in the correct place
