+++ LadderApp/LadderApp.csproj
@@ -70,6 +70,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>MicrocontrollersBaseCodeFilesResource.resx</DependentUpon>
     </Compile>
+    <Compile Include="Services\ExecutableGeneratorServices.cs" />
     <Compile Include="Services\LadderSimulatorServices.cs" />
     <Compile Include="Services\LadderVerificationServices.cs" />
     <Compile Include="Services\Text2OpCodeServices.cs" />
@@ -137,7 +138,7 @@
     <Compile Include="Model\Pin.cs" />
     <Compile Include="VisualComponents\VisualLine.cs">
     </Compile>
-    <Compile Include="Services\MSP430IntegrationServices.cs" />
+    <Compile Include="Services\MicIntegrationServices.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Model\LadderProgram.cs" />
     <Compile Include="VisualComponents\VisualProgram.cs" />
